Add tarball textconv
Create a textconv that displays both the file contents and file metadata
(like mtime, permissions, etc.) to allow for easy comparison of tarballs.
This mechanism only works when the tarball is generated with the same file
order each time -- when using tar directly, this is less likely, but when
using tools like git-archive, this is reliable.