Create external diff to compare tarballs
Tarballs can have similar or identical files but, due to a difference in
file ordering, the textconv may not show that the contents match.  To work
with different file orders, we have to temporarily store the file content,
either in memory or on disk, until we reach a file with a similar name or
content.  This makes comparison much more complicated.

Fortunately, Git already knows how to identify files with similar names or
content.  With some clever use of Git plumbing, we can efficiently extract
files into a temporary Git repository and use Git commands to perform the
diff.  While we still must write to the disk, only compressed blobs are
written - there is no need for a fully uncompressed working copy.

While this proof of concept takes longer than the textconv for tarballs,
it creates more accurate results.  This will enable users to hide the fact
that the contents are stored in a tarball and return results as though the
tarball was extracted into the working copy.