ext-diff/tarball: Improve performance and features
Use libgit2 for interfacing with Git.  This brings significant performance
improvements and enables more advanced features.  However, this change
requires an updated version of Git::Raw that contains a local patch.

The main performance improvement comes from removing the two fork calls
for each file being checked in from the archive.  This improved run times
by an order of magnitude.  In fact, the external diff tool for tarballs is
now faster than the original textconv tool.

Additionally, moving to libgit2 allows us to store all of the data in the
local repository, instead of creating a second, temporary repository that
we discard.  With libgit2 we can write an index in memory without touching
disk, so the local repository - as far as the user is concerned - remains
unchanged.  Of course, the generated trees and blobs are stored in the
repository on disk; however, they are dangling and unseen by most git
commands and they will eventually get cleaned up by git-gc.

Since the generated trees & blobs are now stored in the repository, we can
use them as a cache.  We do this by storing the generated root tree hash
in a git-note that is associated with the source tarball.  Then, if we
process that tarball again, we simply grab the cached value (if it still
exists in the repository).  This caching further improves performance up
to 40%; and it is especially noticeable when walking through the history
with git-log.

Keeping all the data in the original repository also allows us to run git
commands with the same environment.  This gives us access to all of the
config values without copying them over.  Because of this, we now fully
support diff color options on the tarballs.