interdiff: Make color detection smarter
Use git-config to get information from the repository on whether colors
should be used, and to figure out which color should be used.  This
creates a more consistent user interface in git-diff.

This change also cleans up the colorless output.  Previously, the colors
would be reset (with "\e[m") at the end of each line, even if color was
not used.  This would add unreadable junk when the output is piped to a
file; so, for clarity, it is no longer printed on colorless output.