interdiff: Support advanced coloring
Pipe the output through a filter that adds color using a simple heuristic
based on the first few characters.  It can make mistakes, like when adding
a line beginning with "++ " or removing a line beginning with "-- " (which
is actually present in Git patch files), but it's usually correct.

The added (optional) color greatly improves the readability of the
external diff output and makes it more consistent with Git's diff.