interdiff: Annotate interdiff headers for clarity
We relied on the lack of color in the diffs produced by interdiff-patch to
indicate this isn't a raw git diff.  However, for runs where the color was
disabled, there is no such indicator.  And color is a bad way to indicate
anyway.  Add an annotation to each line to show this is not a traditional
Git diff, but rather an interdiff.