textconv/tarball: Show setuid, setgid, & sticky bits
When displaying the permissions access information (the mode), show the
setuid, setgid, & sticky bits in the text form.  Previously, one had to
notice these bits in the octal output, as the 10 character mode string did
not support these values.

This change makes these (likely rare) conditions more obvious and improves
parity with stat(1), the inspiration for the output format.