textconv/tarball: Improve file type handling
Add the GNU long extensions to the list of recognized types, as they are
well understood at this point.  This may help the uninitiated find more
information when it comes up.

Additionally, ensure we always report a value for the name, even if we do
not have an idea what it is.  The stat(1)-like header already does this,
so this change brings the raw header into parity.