ext-diff/tarball: Propagate and support -R
Git reverses the order of the files it passes to this tool when `-R` is
in effect, nearly makes the output correct.  However, the header does
not reverse `a/` & `b/` as it would otherwise, e.g. `diff --git b/ a/`.

To match the correct ordering, we need to un-reverse the extracted trees
and then propagate the `-R` option.  Additionally, the overview needs to
swap the src- and dst-prefixes to properly mimic how git-diff would show
binary files.

Note: This change means that `-R` must now appear by itself in the
calling command; combined options, like `-bR` will not be propagated.
This will be addressed in a future change.