ext-diff/tarball: Convert long options to use hash
Adopt the option propagation hash pioneered for the short options in the
previous commit.  It also adds support for `--option value` in addition
to `--option=value`.  If the former is not supported by Git, the parent
process (e.g. git-diff) will die before this tool is invoked.

As an added bonus, it now becomes much easier to add or remove options
from the propagation list at runtime.  Processing is likely faster, too.