ext-diff/tarball: Audit diff options
Add every diff option available (as of Git v2.28.0) to the option lists,
even if they are unsafe to propagate.  If they are unsafe, add a comment
explaining why it was excluded (making future audits easier to perform).
The options were re-ordered as necessary to match the ordering in the
documentation, also making future audits easier.

As part of this audit, several new options are now propagated; they
include:
* Options that are implied and are therefore (usually) unnecessary
* Options that took a value in a second argument (recently supported)
* Options that have been added to Git since the tool was authored