ext-diff/tarball: Stop caching unindexed files
Files in the working copy that are unindexed (i.e. they are modified)
are passed with a SHA-1 of all zeros.  This means that all unindexed
files will have the same SHA-1 hash; caching on these will inevitably
(and in short order) lead to incorrect results when running a simple
`git diff`.

Users should clear their cache index with `git update-ref -d <ref>` in
order to remove any existing caches (use the value of `--cache-ref` for
`<ref>`).