ext-diff/tarball: Change default cache-ref value
With the recent cache index formatting change, now is an opportune time
to change the location of the git-notes ref.  Doing this will be as if
the user cleared the cache themselves.

This also lets us be more explicit about the purpose of this ref.  By
namespacing it with `/cache/`, we make clear that this ref can safely
be deleted by the user without data loss.