ext-diff/tarball: Change format of cached index
It turns out the \0 delimiters did not even work, as Git::Raw::Notes was
stripping them from the file, so this only worked when there was only a
single entry.  Transition to a line-feed (newline) delimited list to get
this working as expected.

Take the opportunity to revise the cache index format to one that can
support changes to the tree contents for an (unchanged) tarball, for a
means to provide format updates in the future without showing stale
data.

With the new format, the file's path is irrelevant, as we build the path
as needed after getting the tree corresponding to the given blob.  This
means a blob only needs one entry, even if it appears in multiple files
in the repository.  Each entry is annotated, however, with a format ID.
The ID string can be any value, as long as it remains consistent.  When
a new extraction method is introduced, we need only change this ID to
avoid using stale cache data.

Users should clear their cache index with `git update-ref -d <ref>` in
order to remove any existing caches (use the value of `--cache-ref` for
`<ref>`).