ext-diff/tarball: Propagate & support --line-prefix
Change the overview display to work with --line-prefix.  This involves
prepending the prefix to all lines we print, but without coloring them.

This is complicated by the annotations, which Git provides with the
prefix already applied.  To highlight only the annotation and not the
prefix, we must strip off the prefix first, print it without formatting
and, finally, apply formatting and print the actual annotation.

Now that the overview correctly handles line prefixes, this commit lifts
the ban on propagating the --line-prefix= option.  The logic at the time
was to clearly show that we did not support the option by having all of
the diff displayed incorrectly, instead of only having a small, easily
missed sliver of it incorrect.