ext-diff/tarball: Add --indent
Give users an easy way to identify when the diff is of the contents
contained within an archive and are, therefore, uneditable directly.
Indenting the diff with a user specified string (or TAB) clearly
highlights this break.

Notably, the overview is not indented, since that refers to the file
actually in the Git repository.  The user can use --no-overview if they
want the entire external diff indented.