ext-diff/tarball: Add --no-full-path
By default, the path of the tarball is prepended to the full path of
each file within the tarball.  This ensures that the embedded diff
clearly shows the contents come from the tarball.  However, this also
causes each and every file to be listed as renamed when the tarball
itself is renamed, which can be rather noisy.  Add an option that
disables the path prepending to avoid this issue.

To ensure the diff of the internal contents of the tarball are not
misconstrued as part of the repository itself, the tool forces the
--indent option when --no-full-path is given.  Moreover, --indent=
cannot be set to the empty string (by its original design).

As a bonus, this reduces the length of the filenames displayed for all
tarballs, which can be useful.