ext-diff/tarball: Clarify function output
Future work will allow users to change this tools behavior with command
line options (given before the arguments provided by Git), which will
undoubtedly lead to confusion over what `get_options()` does.  Rename
the function and comment on it to clarify that this collects the options
to use on git-diff.