ext-diff/tarball: Move global vars to config hash
Get rid of the global variables for modifying behavior and replace them
with a single configuration hash that gets passed down the call stack.

This also separates the caching mechanism into two separate variables, a
toggle and the ref to use when the toggle is enabled.  This allows the
person editing the file to disable the cache without removing the cache
ref, making it easier to reenable to the same ref as before.