ext-diff/tarball: Allow user to amend diff options
Add options to let the user prepend or append to the propagated options
passed to git-diff.  The former allow users to set defaults that can be
overridden by options on the parent diff, while the latter allows any
option to be forced.

Combined with --no-propagate-diff-options, either of these commands can
be used to specify a different set of diff options that always get used.