ext-diff/tarball: Add --munge-components
Most tarballs will include all of their contents in a single top-level
directory; that way, when the contents are extracted, they aren't strewn
about the current directory (making it difficult to identify and remove
them).  With many versioned packages, this directory will include the
version in the name, so it changes each and every time the tarball is
packaged; thus, each and every file is "renamed" according to Git.

To avoid the rename noise, and to improve file break detection, provide
an option akin to tar's --strip-components option, --munge-components.
This allows the differences in the top level directory(ies) to be
hidden, as they are irrelevant.  The data is munged, not stripped, to
ensure a record of the alteration is clearly present.  Indeed, munging
too many directory levels can cause filename collisions to occur, such
that the diff is not correctly represented, so this record provides a
means to help with troubleshooting.