ext-diff/tarball: Propagate --stat, --raw on indent
When the --indent option is given to the tool, propagate `--raw`,
`--raw-with-patch`, and the `--stat` family of options.  Note: as
both --raw and --stat replace the patch output, this tool will not
be run without also giving `-p`, or an equivalent option.

These options are only propagated when --indent is given or implied, as
raw or stat information can appear in the middle of patch output, which
can break diff parsers and creates unclear separation visually.  Having
an indent clearly isolates the sub-diff and also allows users to prefix
the information with something a diff parser might ignore (like context
lines).