reword! ext-diff/tarball: Add --no-show-content
Define a more extensible cache index format naming scheme, and provide
an option to change the format being used.  While the --no-show-content
option is not particularly useful by itself, as all tarballs will end up
with the empty tree (making them all identical), it provides a mechanism
for testing differing formats in the cache index.

Future options for selecting the format can benefit from being able to
exclude file content as it may be able to limit noise when looking for
specific tarball differences.  That said, this is not expected to be a
frequently used option.