ext-diff/tarball: Propagate diff filter
Provide a means to limit the type of changes that are presented in the
sub-diff generated by this tool.  Of course, for this tool to be run in
the first place, the tarball itself needs to be included in the filter.
As such, it is recommended that users explicitly include the change type
of the tarball in addition to what they are actually looking for.