ext-diff/tarball: Propagate diff search options
Allow the use of `-S` and `-G` search options for limiting the scope of
the diff.  For this to work, those same string must appear verbatim in
the tarball so Git will select the tarball in the first place; this
cannot happen on compressed tarballs.  As such, this change also
disables support for processing compressed tarballs.