merge/sqlite3: Use configured mergetool after conv
After doing the textconv, this tool runs the _actual_ mergetool that
lets the user merge the text.  Previously, the tool used was hardcoded.
This changes it to use whatever tool is set up by the user, either in
git-config or through an environmental variable the user can set.

With this change, this tool is now ready for wider adoption, as users
can customize it to meet their needs without modifying its source code.