ext-diff/tarball: Isolate and document format ID
Give the cache format identifier its own function, complete with a set
of guidelines on how to modify it, in addition to a detailed description
on how it will be used.  This prepares the tool for additional changes
in output that can be safely cached independently of the default format.